Continuing to decouple
This is an update where I am continuing to decouple the classes.

Next, I’m going to need to take pics away from the GameBoard class. And
it’s really not necessary to have an image as part of the Cell class.
But I’ll leave it there for a while anyway.